% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStats.R
\name{calcSandT}
\alias{calcSandT}
\title{Calculate 2nd and 3rd moment tensors}
\usage{
calcSandT(Y)
}
\arguments{
\item{Y}{n x p matrix of observation}
}
\value{
A list of sMat and tMat which are the 2nd and 3rd order moment tensors
}
\description{
Calculate 2nd and 3rd moment tensors
}
