% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCycle.R
\name{constructCycle}
\alias{constructCycle}
\title{Estimate edges in a root cycle}
\usage{
constructCycle(sMat, tMat)
}
\arguments{
\item{sMat}{the covariance matrix for the variables in a root cycle.
If the cycle is not a root cycle, it should be the covariance of the variables after adjusting
for nodes in preceeding layers}

\item{tMat}{the third moment tensor for the variables in a root cycle.
If the cycle is not a root cycle, it should be the third moments of the variables after adjusting
for nodes in preceeding layers}
}
\value{
\itemize{
\item adjMat the estimated adjacency matrix where adjMat[i,j] == 1 indicates the edge i -> j
\item Lambda the estimated edge weights where Lambda[i,j] indicates the linear coefficient of i onto j
}
}
\description{
Estimate edges in a root cycle
}
