% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{cycleChain}
\alias{cycleChain}
\title{Generate Lambda which is a chain of cycles}
\usage{
cycleChain(
  p,
  cycleSize = 2,
  lowEdge = 0.5,
  highEdge = 0.9,
  posNeg = T,
  parentProb = 1,
  uniqueTop = T
)
}
\arguments{
\item{p}{the number of variables}

\item{cycleSize}{the size of each cycle}

\item{lowEdge}{the lower bound on the absolute value of each edgeweight}

\item{highEdge}{the upper bound on the absolute value of each edgeweight}

\item{posNeg}{whether the edgeweights can be positive and negative}

\item{parentProb}{the probability of an edge between an ancestor and descendant not in the same cycle}

\item{uniqueTop}{Whether to enforce a unique topological ordering of cycles by adding an edge from one cycle
to the next with probability 1}
}
\value{
a p x p matrix with edgeweight
}
\description{
Generate Lambda which is a chain of cycles
}
