% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djcGetOrderNew.R
\name{djc_recurNew}
\alias{djc_recurNew}
\title{a root singleton) or of length greater than 1 (indicating a root cycle)
 newY: an n x p matrix which contains Y for the remaining unordered nodes}
\usage{
djc_recurNew(
  Y,
  alpha2 = 0.01,
  alpha3 = 0.01,
  alphaR = 0.01,
  methodRoot = "indDelta",
  methodDet2 = "indDelta",
  rescaleData = T,
  pvalAdjMethod,
  methodPR,
  sigmaPop = NULL,
  rescaleResid = T
)
}
\description{
a root singleton) or of length greater than 1 (indicating a root cycle)
 newY: an n x p matrix which contains Y for the remaining unordered nodes
}
