% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djcGetEdges.R
\name{djcGetEdges}
\alias{djcGetEdges}
\title{Estimate edges when given a topological ordering fo cyclic graphs with disjoint cycles}
\usage{
djcGetEdges(
  topOrdering,
  Y,
  alpha,
  verbose = F,
  rescaleData = T,
  pvalAdjMethod = "holm"
)
}
\arguments{
\item{topOrdering}{a list of lists containing the topological layers estimated by djcGetOrder}

\item{Y}{n x p matrix of observations}

\item{alpha}{the level for the hypothesis test when pruning parent edges}

\item{verbose}{print out results for each step}

\item{rescaleData}{whether to rescale data after each step}

\item{pvalAdjMethod}{the multiple testing adjustment procedure}
}
\value{
\itemize{
\item adjMat the estimated adjacency matrix where adjMat[i,j] == 1 indicates the edge i -> j
\item Lambda the estimated edge weights where Lambda[i,j] indicates the linear coefficient of i onto j
\item pvals_adj the adjusted p-values for potential parents to be pruned. If the value is -1 this indicates
      that there was no test for the parent in this stage. This is because it is part of a cycle so the
      edge was forced to exist or not exist based on the greedy algorithm or because one does not
      preceed the other in the topological layers.
}
}
\description{
Estimate edges when given a topological ordering fo cyclic graphs with disjoint cycles
}
