% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djcGetOrderNew.R
\name{djcGetOrderNew}
\alias{djcGetOrderNew}
\title{Estimate topological ordering for cyclic graphs with disjoint cycles}
\usage{
djcGetOrderNew(
  Y,
  alpha2 = 0.01,
  alpha3 = 0.01,
  alphaR = 0.01,
  methodRoot = "indDelta",
  methodDet2 = "indDelta",
  rescaleData = T,
  verbose = F,
  pvalAdjMethod = "holm",
  methodPR = "infFunc",
  sigmaPop = NULL,
  rescaleResid = T
)
}
\arguments{
\item{Y}{n x p matrix of observations}

\item{alpha2}{the level for the hypothesis test when determining singleton roots}

\item{alpha3}{the level for the hypothesis test when determining}

\item{alphaR}{the level for the hypothesis test when}

\item{methodRoot}{which method to use for testing singleton roots;
i.e., for fixed u \eqn{H_0: det_{u,v}^{2x2} == 0} for v!=u. Options are indDelta which tests each value individually using
asymptotic normality and the adjusting for multiple testing with holm or jointDelta which uses asymptotic normality of entire vector
 and does a single chi-squared test}

\item{methodDet2}{which method to use for testing \eqn{H_0: det_{u,v}^{2x2} != 0} or \eqn{det_{v,u}^{2x2} != 0}
Options are indDelta which tests each value individually using
 asymptotic normality and the adjusting for multiple testing with holm or jointDelta which tests whether the product of the two quantities
 are non-zero and uses asymptotic normality}

\item{rescaleData}{whether to rescale data after each step}

\item{verbose}{print out results for each step}
}
\value{
A list of lists where each internal list contains vectors of nodes which are either of length 1 (indicating
a root singleton) or of length greater than 1 (indicating a root cycle)
}
\description{
Estimate topological ordering for cyclic graphs with disjoint cycles
}
