% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneRoots.R
\name{pruneHelperRootCycle}
\alias{pruneHelperRootCycle}
\title{Helper function used to select a root cycle from the set C}
\usage{
pruneHelperRootCycle(C, D, Y, methodPR = "infFunc", rescaleResid = T)
}
\arguments{
\item{C}{the clique of interest}

\item{D}{all other nodes}

\item{Y}{the n x p data matrix}

\item{methodPR}{the method to be used: chisq, infFunc, or naive}
}
\value{
\itemize{
\item B the matrix of edge weights
\item Y the n x p data
\item errs the error realizations
\item scales the variance of the errors
\item mu the intercept terms
}
}
\description{
Helper function used to select a root cycle from the set C
}
