% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCycle.R
\name{pruneHelperSingle}
\alias{pruneHelperSingle}
\title{Helper function to test existence of edges from nodes in preceeding layer
used when D is cycle}
\usage{
pruneHelperSingle(C, D, Y, lambda_CD, pr = "infFunc")
}
\arguments{
\item{C}{the set of nodes in the preceeding layer}

\item{D}{the set of nodes in the current unit (must be cycle)}

\item{dInd}{the specific index in D to test}
}
\value{
the pvalues (unadjusted for multiple testing) for any edge to be tested. If the value is -1 this indicates
      that there was no test for the parent in this stage. This is because it is part of a cycle so the
      edge was forced to exist or not exist based on the greedy algorithm or because one does not
      preceed the other in the topological layers.
}
\description{
Helper function to test existence of edges from nodes in preceeding layer
used when D is cycle
}
