% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructCycle.R
\name{pruneParents}
\alias{pruneParents}
\title{Test existence of edges from nodes in preceeding layer}
\usage{
pruneParents(C, D, sMat, tMat, Y)
}
\arguments{
\item{C}{the set of nodes in the preceeding layer}

\item{D}{the set of nodes in the current unit (can be a singleton or cycle)}
}
\value{
\itemize{
\item adjMat the estimated adjacency matrix where adjMat[i,j] == 1 indicates the edge i -> j
\item Lambda the estimated edge weights where Lambda[i,j] indicates the linear coefficient of i onto j
\item pvals the pvalues (unadjusted for multiple testing) for any edge to be tested. If the value is -1 this indicates
      that there was no test for the parent in this stage. This is because it is part of a cycle so the
      edge was forced to exist or not exist based on the greedy algorithm or because one does not
      preceed the other in the topological layers.
}
}
\description{
Test existence of edges from nodes in preceeding layer
}
