% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rLSEM.R
\name{rLSEM}
\alias{rLSEM}
\title{Sample data from an additive noise model}
\usage{
rLSEM(
  p,
  n,
  lowScale = 1,
  highScale = 1,
  lowEdge = 0.3,
  highEdge = 1,
  dist = "gauss",
  AdjMat = NULL,
  LambdaIn = NULL,
  scalesIn = NULL,
  posAndNeg = T
)
}
\arguments{
\item{p}{number of variables}

\item{n}{number of observations}

\item{lowScale}{lower bound on variance of error terms}

\item{highScale}{upper bound on variance of error terms}

\item{lowEdge}{lower bound on edgeWeights}

\item{highEdge}{upper bound on edgeWeights}

\item{dist}{the distribution of the error terms. Choices are: "gauss", "unif", "lognorm", "gamma", "weibull", "laplace", "mixed"}

\item{AdjMat}{an adjacency matrix which can be passed in to specify the edges instead of randomly selecting a graph.
If A[i,j] == 1, this indicates i --> j}

\item{LambdaIn}{a matrix of edges which can be passed in instead of randomly drawing edges}

\item{scalesIn}{a vector of error variances which can be passed in instead of randomly drawing}

\item{posAndNeg}{whether edges should be both positive and negative or only positive}
}
\value{
\itemize{
\item B the matrix of edge weights
\item Y the n x p data
\item errs the error realizations
\item scales the variance of the errors
\item mu the intercept terms
}
}
\description{
Sample data from an additive noise model
}
