% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{rMixNorm}
\alias{rMixNorm}
\title{Generate data from a mixture of normals}
\usage{
rMixNorm(n, mu, sd, piMix)
}
\arguments{
\item{n}{the sample size}

\item{mu}{each component is centered at +/- mu}

\item{sd}{the standard deviation of each component}

\item{piMix}{the proportion of the sample for one of the components. This is randomly assigned to
either the positive or negative component so the skew may be right or left.}
}
\value{
a vector of length n
}
\description{
Generate data from a mixture of normals
}
